requirements.txt
boto3==1.34.122
gensim==4.3.2
nltk==3.8.1
numpy==2.0.1
pandas==2.0.3
polars==1.4.1
pyarrow==16.1.0
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
Requests==2.32.3
spacy==3.7.4
scipy==1.11.2
SQLAlchemy==2.0.27
tqdm==4.65.0
